Eliminar carrusel pasado y corregir imports
