+++ DREAM Lab Frontend/src/Home/HomePage.jsx
@@ -1,13 +1,11 @@
-import React, { useState, useEffect } from "react";
+import { useState, useEffect } from "react";
 import ImageSlider from './ImageSlider'
 import SpeechBotCard from './SpeechBotCard'
 import '../App.css'
 import GlassCard from '../components/general/glass-card'
-import RecommendationsCarousel from './RecommendationsCarousel'
 import "./RecommendationsCarousel.css"
 import UserAvatar from '../components/general/UserAvatar'
-import NCarruselRecomendaciones from './NCarruselRecomendaciones'
-import Typewriter from "./Typewriter";
+import RecommendationsCarousel from './RecommendationsCarousel'
 
 const OPTIONS = { dragFree: true, loop: true }
 
@@ -89,7 +87,7 @@ function HomePage() {
             <SpeechBotCard width='100%' height='25rem' onProcessedText={handleProcessedText} />
 
             {showRecommendations && (
-                <NCarruselRecomendaciones data={data} activeSlide={parseInt(Math.floor(data.length / 2))} />
+                <RecommendationsCarousel data={data} activeSlide={parseInt(Math.floor(data.length / 2))} />
             )}
 
             <br />
@@ -100,8 +98,6 @@ function HomePage() {
             <ImageSlider images={IMAGES} options={OPTIONS} />
             <br />
             <ImageSlider images={IMAGES} options={OPTIONS} />
-
-            <Typewriter />
         </>
     )
 }
