+++ DREAM Lab Frontend/src/Home/RecommendationsCarousel.css
@@ -1,22 +1,81 @@
-@import url('https://fonts.googleapis.com/css2?family=Karla:ital,wght@0,200..800;1,200..800&display=swap');
+.slideC {
+	position: relative;
+	perspective: 1000px;
+	transform-style: preserve-3d;
+	width: 362px;
+	height: 272px;
+	margin: 0 auto;
+}
 
-.imageCarousel {
-	user-select: none;
+.slide {
+	width: 362px;
+	height: 272px;
+	transition: transform 500ms ease 0s, opacity 500ms ease 0s,
+		visibility 500ms ease 0s;
+	position: absolute;
+	top: 0;
+	border-radius: 12px;
+}
+
+/* 3D reflection effect */
+.reflection {
+	position: absolute;
 	width: 100%;
-	height: 100%;
-	object-fit: cover;
-	border-radius: 2rem;
+	height: 60px;
+	bottom: -60px;
+	border-radius: 12px;
+	transition: transform 500ms ease 0s, opacity 500ms ease 0s,
+		visibility 500ms ease 0s;
 }
 
-.carousel-container {
-    display: flex;
-    justify-content: center;
-}
-
-.processed-transcript-container {
+.slide {
 	display: flex;
-    justify-content: center;
-	font-family: 'Karla', sans-serif;
-	color: white;
-	font-size: large;
+	justify-content: left;
+	align-items: left;
 }
+
+.sliderContent {
+	display: flex;
+	flex-direction: column;
+	color: white;
+	padding: 30px;
+	align-items: flex-start;
+	font-family: "Lucida Sans", "Lucida Sans Regular", "Lucida Grande",
+		"Lucida Sans Unicode", Geneva, Verdana, sans-serif;
+
+}
+
+.sliderContent h2 {
+	margin: 16px 0;
+}
+
+.sliderContent p {
+	margin: 0;
+	margin-bottom: 16px;
+}
+
+.center {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	flex-direction: column;
+	width: 100%;
+	height: 100vh;
+}
+
+.btns {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	padding-top: 100px;
+}
+
+.btns .btn {
+	cursor: pointer;
+}
+
+.btns .btn:last-child {
+	margin-left: 40px;
+}
+
+/* normal css  */
